fixed int overflow in BlockChainData

+++ wavesplatform/src/main/java/com/wavesplatform/sdk/model/response/node/BlockChainData.kt
@@ -13,5 +13,5 @@ data class BlockChainData(
     @SerializedName("type")
     var type: String,
     @SerializedName("value")
-    var value: Int
+    var value: Long
 )
